requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
redis==5.2.0
APScheduler==3.10.4
python-dotenv==1.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

